# control-plane/internal/apispec/openapi.yaml
openapi: 3.1.0
info:
  title: Hive API
  version: 0.5.0
  description: 'Hive is a Swarm-native platform control plane. This spec is the single

    source of truth for all API contracts. Every request body, response, and

    error shape is defined here.

    '
servers:
- url: /api/v1
  description: Control plane API v1
components:
  schemas:
    ErrorResponse:
      description: ErrorResponse schema.
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
      example:
        code: bad_request
        message: invalid payload
    StatusResponse:
      description: StatusResponse schema.
      type: object
      additionalProperties: false
      required:
      - status
      properties:
        status:
          type: string
          description: Current status.
      example:
        status: ok
...
    IdResponse:
      description: IdResponse schema.
      type: object
      additionalProperties: false
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
...
    Registry:
      description: Registry schema.
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - url
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
        name:
          type: string
          description: Display name.
        url:
          type: string
          description: URL address.
        username:
          type: string
          description: Username for authentication.
        secretName:
          type: string
          description: Secret name.
        isDefault:
          type: boolean
          description: Is default.
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601).
    CreateRegistryRequest:
      description: CreateRegistryRequest schema.
      type: object
      additionalProperties: false
      required:
      - name
      - url
      properties:
        name:
          type: string
          description: Display name.
        url:
          type: string
          description: URL address.
        username:
          type: string
          description: Username for authentication.
        secretName:
          type: string
          description: Secret name.
        isDefault:
          type: boolean
          default: false
          description: Is default.
      example:
        name: Docker Hub
        url: https://index.docker.io/v1/
        username: dockeruser
    UpdateRegistryRequest:
      description: UpdateRegistryRequest schema.
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          description: Display name.
        url:
          type: string
          description: URL address.
        username:
          type: string
          description: Username for authentication.
        secretName:
          type: string
          description: Secret name.
        isDefault:
          type: boolean
          description: Is default.
      example:
        name: Docker Hub Pro
...
    Redirect:
      description: HTTP redirect rule for domain routing
      type: object
      required:
      - id
      - domainId
      - path
      - target
      - statusCode
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Redirect ID
        domainId:
          type: string
          format: uuid
          description: Domain ID
        path:
          type: string
          description: Source path pattern
        target:
          type: string
          description: Target URL or path
        statusCode:
          type: integer
          enum:
          - 301
          - 302
          - 307
          - 308
          description: HTTP redirect status code
        permanent:
          type: boolean
          description: Whether this is a permanent redirect
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        domainId: 550e8400-e29b-41d4-a716-446655440001
        path: /old-path
        target: /new-path
        statusCode: 301
        permanent: true
        createdAt: '2026-01-01T00:00:00Z'
    RedirectList:
      description: List of redirects
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Redirect'
    CreateRedirectRequest:
      description: Create redirect request
      type: object
      required:
      - domainId
      - path
      - target
      - statusCode
      properties:
        domainId:
          type: string
          format: uuid
          description: Domain ID
        path:
          type: string
          description: Source path pattern
        target:
          type: string
          description: Target URL or path
        statusCode:
          type: integer
          enum:
          - 301
          - 302
          - 307
          - 308
          description: HTTP redirect status code
        permanent:
          type: boolean
          description: Whether this is a permanent redirect
      example:
        domainId: 550e8400-e29b-41d4-a716-446655440001
        path: /old-path
        target: /new-path
        statusCode: 301
        permanent: true
...
  parameters:
    IdParam:
      description: IdParam schema.
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      description: bearerAuth schema.
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /registries:
    get:
      operationId: listRegistries
      summary: List registries
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Registry list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Registry'
      description: Retrieve registries
    post:
      operationId: createRegistry
      summary: Create registry
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistryRequest'
      responses:
        '201':
          description: Registry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Create registries
  /registries/{id}:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getRegistry
      summary: Get registry
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registry'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Retrieve registries id
    put:
      operationId: updateRegistry
      summary: Update registry
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegistryRequest'
      responses:
        '200':
          description: Registry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Update registries id
    delete:
      operationId: deleteRegistry
      summary: Delete registry
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Registry deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Delete registries id
  /registries/{id}/test:
    parameters:
    - $ref: '#/components/parameters/IdParam'
    post:
      operationId: testRegistry
      summary: Test registry connectivity
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Registry test passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Create registries id test
      requestBody:
        description: Empty request body.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
...
  /redirects:
    get:
      operationId: listRedirects
      summary: List redirects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectList'
      description: List all redirect rules
      security:
      - bearerAuth: []
    post:
      operationId: createRedirect
      summary: Create redirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRedirectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
      description: Create a new redirect rule
      security:
      - bearerAuth: []
  /redirects/{id}:
    get:
      operationId: getRedirect
      summary: Get redirect
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redirect'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Get redirect by ID
      security:
      - bearerAuth: []
    put:
      operationId: updateRedirect
      summary: Update redirect
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRedirectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Update redirect rule
      security:
      - bearerAuth: []
    delete:
      operationId: deleteRedirect
      summary: Delete redirect
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                description: No response body.
        '404':
          $ref: '#/components/responses/NotFound'
      description: Delete redirect rule
      security:
      - bearerAuth: []
...